Move mapToData function out from usePerformance for reuse
